requirements.txt
aniso8601==9.0.1
apispec==5.1.1
async-timeout==4.0.2
attrs==21.4.0
Cerberus==1.3.4
click==8.1.2
Deprecated==1.2.13
Flask==2.1.1
Flask-And-Redis==1.0.0
flask-apispec==0.11.1
Flask-Caching==1.10.1
flask-expects-json==1.7.0
Flask-HTTPAuth==4.5.0
Flask-Inputs==0.3.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing_extensions==4.1.1
webargs==8.1.0
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
zipp==3.8.0
